requirements.txt
wheel
flask
gevent
pymssql
flask-login
#Flask-WTF
#flask-cors
python-docx
